requirements.txt
Flask
flask-sqlalchemy
psycopg2-binary
PyYAML